Extracts the show number button to a separate file
